linux: fix real portability bugs surfaced by the first Linux build attempt
Found by actually building the Linux target (meson/dxvk-native aside,
Misc/FileIO/Image are buildable now and were used to drive this):

- Misc/tools.h: remaining __asm ("_asm", missed by the earlier audit's
  "__asm\b" pattern) and __int64 guarded for non-MSVC (Sign<int/short>,
  MemSetDWord, Float2Int, Min/Max<float>, GetCPUID -- the last unused
  anywhere in the codebase, verified).
- Misc/basic2.h, Misc/BasicShare.h: GCC's two-phase template lookup
  needs explicit `this->`/`typename` for names MSVC resolved more
  loosely (BASIC_PTR_DECLARE's comparison operators, EraseInvalidRefs,
  CBasicShare's iterators). Syntax only, no behavior change.
- Misc/EventsBase.cpp: (int)&type_info truncated a pointer to 32 bits --
  harmless on Win32 (4-byte pointers) but a real bug on 64-bit Linux
  (colliding hash keys). Widened the event-id map's key to intptr_t.
- Misc/RandomGen.cpp: FillRandRsl's entropy source (scan C:\*.* for a
  random file, read bytes from it) has no Windows-independent meaning;
  replaced with /dev/urandom on non-MSVC, which is a strictly better
  entropy source anyway. `unsigned _int8` -> `unsigned char` (portable
  either way).
- third_party/zlib/zconf.h: ZEXTERN/ZEXPORTVA unconditionally expanded
  to __declspec(...)/__cdecl regardless of ZLIB_DLL -- added the missing
  _WIN32 guard this vendored copy never had.
- third_party/libpng/png.h: backslash #include, same class of fix as
  the repo-wide one (this file lives under third_party/, which that
  pass skipped).
- Misc/PlatformCompat.h: added QueryPerformanceCounter/Frequency
  (CLOCK_MONOTONIC-backed, HPTimer.cpp's calibration window) and
  IsBadReadPtr (msync-based unmapped-page probe, Basic2.cpp's
  use-after-free guard during save/load teardown).

+++ Misc/EventsBase.cpp
@@ -14,19 +14,21 @@ static struct SExecutionTracker
 } tracker;
 //
 typedef vector<IEventRegister*> CCallInfoHash;
-static unordered_map< int, CCallInfoHash > *pEventHandlers = 0;
+// key = the type_info's address, reused as a per-type identity. int held it fine when
+// pointers were 4 bytes (Win32); truncates on 64-bit Linux, so widen to intptr_t there.
+static unordered_map< intptr_t, CCallInfoHash > *pEventHandlers = 0;
 static int nEventHandlersCount = 0;
 ////////////////////////////////////////////////////////////////////////////////////////////////////
-inline unordered_map< int, CCallInfoHash > &GetEventHandlers()
+inline unordered_map< intptr_t, CCallInfoHash > &GetEventHandlers()
 {
 	if ( pEventHandlers == 0 )
-		pEventHandlers = new unordered_map< int, CCallInfoHash >();
+		pEventHandlers = new unordered_map< intptr_t, CCallInfoHash >();
 	return *pEventHandlers;
 }
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 void ThrowEventInner( const type_info &eventID, const void *pStuff )
 {
-	int nEventID = (int)&eventID;
+	intptr_t nEventID = (intptr_t)&eventID;
 	CCallInfoHash &handlers = GetEventHandlers()[ nEventID ];
 	for ( CCallInfoHash::iterator i = handlers.begin(); i != handlers.end(); ++i )
 		(*i)->Call( pStuff );
@@ -34,14 +36,14 @@ void ThrowEventInner( const type_info &eventID, const void *pStuff )
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 void RegisterEventHandler( IEventRegister *pReg, const type_info &eventID )
 {
-	int nEventID = (int)&eventID;
+	intptr_t nEventID = (intptr_t)&eventID;
 	GetEventHandlers()[ nEventID ].push_back( pReg );
 	++nEventHandlersCount;
 }
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 void UnregisterEventHandler( IEventRegister *pReg, const type_info &eventID )
 {
-	int nEventID = (int)&eventID;
+	intptr_t nEventID = (intptr_t)&eventID;
 	vector<IEventRegister*> &handlers = GetEventHandlers()[ nEventID ];
 	vector<IEventRegister*>::iterator i = find( handlers.begin(), handlers.end(), pReg );
 	if ( i != handlers.end() )
